FocalAdjust
FocalAdjust allows you to enter focal length of a camera lens (ie 28mm) along with the crop factor (ie 1.5x) to get the new focal length (42mm). This works with all focal lengths and crop factors (as it is just simple math after all). It may not be a very ground breaking or even useful tool, but I really wanted to write a program so this is what I went with. 

I'm working on adding a GUI to the program, but I'm currently having issues with Tkinter so it'll likely come at a later date. 

ALSO, this program was greatly written by ChatGPT. I made additions and changed a few things to fit more what I was wanting, but I gotta give credit where it's due... Thanks AI!